Improved the responsiveness of the website
Added img-responsive classes to the project images.  Saw big
improvement.  Also added min-height to paragraph div’s.  It helped…to a
point.  It is either I make bigger blank div or let it adjust to the
text…I think for balanced visual, I have to give up after 250px, and
edit the text accordingly (cut the word count down to fit the div).